+++ js/form/load-images.js
@@ -1,0 +1,1 @@
+//Загрузка фотографий
