+++ js/form/pin-location.js
@@ -1,0 +1,1 @@
+//Отметка локации объявления на карте
