+++ js/form/validation.js
@@ -1,0 +1,1 @@
+//Обработчик событий (success/error) при публикации объявления
