+++ js/map/filter.js
@@ -1,0 +1,1 @@
+//Фильтрация схожих объявлений на карте
