+++ js/map/load-ads.js
@@ -1,0 +1,1 @@
+//Загрузка на карту схожих объявлений
