+++ js/map/map.js
@@ -1,0 +1,1 @@
+//Работа с картой
