+++ js/map/popup-ad.js
@@ -1,0 +1,1 @@
+//Просмотр карточки схожего объявления
